refactor(Aside): SSR로 렌더링방식 변경, useEffect 삭제
